tests: skip kernel test when lttng driver not install on target rootfs
The current tests will run both userspace and kernel testing.
Some of use cases only use lttng for one kind of tracing on
an embedded device (e.g. userspace).
If the lttng modules(.ko files) is not present during the test,
it would end up with lots of failing.

Adding a function in utils.sh to detect whether lttng modules are
installed to target rootfs.
Also applied this function to test scripts, if modules not exist,
skip the test case instead of end up whole testing.

Signed-off-by: Xiangyu Chen <xiangyu.chen@windriver.com>